Updated issue I definetley didnt cause

+++ cogs/moderation/kick.py
@@ -15,30 +15,33 @@ class kick(commands.Cog):
     async def kick(self, ctx, member: discord.Member, *, reason = "No reason specified"):
         user = ctx.author
         if any(role.id in rolelist for role in user.roles):
-            await ctx.message.delete()
-            await self.new_warn_member(member)
-            channel = self.client.get_channel(933768368970932254)
-            try:
-                await member.send(f"You were kicked from the PC Creater server for:\n" + reason)
+            if any(role.id in rolelist for role in member.roles):
+                await ctx.message.delete()
+                await self.new_warn_member(member)
+                channel = self.client.get_channel(933768368970932254)
+                try:
+                    await member.send(f"You were kicked from the PC Creater server for:\n" + reason)
 
-                await member.kick(reason=reason)
-                await ctx.send(f"Kicked {member.mention}", delete_after=10)
+                    await member.kick(reason=reason)
+                    await ctx.send(f"Kicked {member.mention}", delete_after=10)
 
-                embed = discord.Embed(title="Kicked", color=13565696)
-                embed.add_field(name="Kicked:", value=f"{member.mention}")
-                embed.add_field(name="Moderator", value=f"{ctx.author.mention}")
-                embed.add_field(name="Reason:", value=reason, inline=False)
-                await channel.send(embed=embed)
-            except:
-                await member.kick(reason=reason)
-                await ctx.send(f"Kicked {member.mention}", delete_after=10)
+                    embed = discord.Embed(title="Kicked", color=13565696)
+                    embed.add_field(name="Kicked:", value=f"{member.mention}")
+                    embed.add_field(name="Moderator", value=f"{ctx.author.mention}")
+                    embed.add_field(name="Reason:", value=reason, inline=False)
+                    await channel.send(embed=embed)
+                except:
+                    await member.kick(reason=reason)
+                    await ctx.send(f"Kicked {member.mention}", delete_after=10)
 
-                embed = discord.Embed(title="Kicked", color=13565696)
-                embed.add_field(name="Kicked:", value=f"{member.mention}")
-                embed.add_field(name="Moderator", value=f"{ctx.author.mention}")
-                embed.add_field(name="Reason:", value=reason, inline=False)
-                await channel.send(embed=embed)
-            await self.update_warns(member, reason)
+                    embed = discord.Embed(title="Kicked", color=13565696)
+                    embed.add_field(name="Kicked:", value=f"{member.mention}")
+                    embed.add_field(name="Moderator", value=f"{ctx.author.mention}")
+                    embed.add_field(name="Reason:", value=reason, inline=False)
+                    await channel.send(embed=embed)
+                await self.update_warns(member, reason)
+            else:
+                await ctx.send(f"Cannot ban a member with Moderator permissions.", delete_after=10)
         else:
             return
